backend/requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1602530305083/work
astor==0.8.1
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1607706694405/work
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1606255099073/work
chardet @ file:///tmp/build/80754af9/chardet_1605303159953/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
cryptography @ file:///tmp/build/80754af9/cryptography_1607635305226/work
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1607969906642/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1597424467487/work
h5py==2.8.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
Markdown @ file:///tmp/build/80754af9/markdown_1605111062842/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools @ file:///tmp/build/80754af9/more-itertools_1605111547926/work
multidict @ file:///tmp/build/80754af9/multidict_1600456400975/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786740831/work
pluggy==0.13.1
protobuf==3.13.0
py @ file:///tmp/build/80754af9/py_1607971587848/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
pytest==6.2.2
pytest-asyncio @ file:///tmp/build/80754af9/pytest-asyncio_1600458979745/work
PyYAML==5.4.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1610483308194/work
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
six @ file:///tmp/build/80754af9/six_1605205313296/work
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
websockets @ file:///home/conda/feedstock_root/build_artifacts/websockets_1602327016732/work
Werkzeug==1.0.1
wrapt==1.12.1
yarl @ file:///tmp/build/80754af9/yarl_1606939952652/work
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
